Refactor test conftest
