+++ tests/python/contrib/test_hexagon/rpc/conftest.py
@@ -14,11 +14,8 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+
 import pytest
-import os
-
-import tvm.testing
-
 
 def pytest_addoption(parser):
     parser.addoption(
@@ -31,52 +28,3 @@ def pytest_addoption(parser):
 @pytest.fixture
 def android_serial_number(request):
     return request.config.getoption("--serial-number")
-
-
-@tvm.testing.fixture
-def tvm_tracker_host():
-    return os.environ["TVM_TRACKER_HOST"]
-
-
-@tvm.testing.fixture
-def tvm_tracker_port():
-    return int(os.environ["TVM_TRACKER_PORT"])
-
-
-def _compose(args, decs):
-    """Helper to apply multiple markers"""
-    if len(args) > 0:
-        f = args[0]
-        for d in reversed(decs):
-            f = d(f)
-        return f
-    return decs
-
-
-def requires_rpc_tracker(*args):
-    """Mark a test as requiring an RPC tracker to exist in
-    the host environment to run."""
-    _requires_rpc_tracker = [
-        *tvm.testing.requires_rpc(),
-        pytest.mark.skipif(
-            os.environ.get("TVM_TRACKER_HOST") == None,
-            reason="Missing environment variable, TVM_TRACKER_HOST",
-        ),
-        pytest.mark.skipif(
-            os.environ.get("TVM_TRACKER_PORT") == None,
-            reason="Missing environment variable, TVM_TRACKER_PORT",
-        ),
-    ]
-
-    return _compose(args, _requires_rpc_tracker)
-
-
-def requires_hexagon_toolchain(*args):
-    _requires_hexagon_toolchain = [
-        pytest.mark.skipif(
-            os.environ.get("HEXAGON_TOOLCHAIN") == None,
-            reason="HEXAGON_TOOLCHAIN environment variable is required to run this test.",
-        ),
-    ]
-
-    return _compose(args, _requires_hexagon_toolchain)
